Revert "Revert "Refactoring""
