+++ src/js/fetch/fetch-films.js
@@ -1,7 +1,6 @@
 import axios from 'axios';
 
-const KEY = '3c2d3d1a4a9318a7ef02a0fdedccb03f';
-const BASE_URL = 'https://api.themoviedb.org/3/search/movie';
+import { KEY, BASE_URL } from './key-url';
 
 export default class FilmsAPI {
   constructor() {
@@ -10,7 +9,7 @@ export default class FilmsAPI {
   }
 
   async fetchFilms(page) {
-    const url = `${BASE_URL}?api_key=${KEY}&query=${this.searchQuery}&page=${page}`;
+    const url = `${BASE_URL}/search/movie?api_key=${KEY}&query=${this.searchQuery}&page=${page}`;
 
     const { data } = await axios.get(url);
