+++ src/js/fetch/fetch-trailer.js
@@ -1,8 +1,6 @@
 import axios from 'axios';
 
-const KEY = '3c2d3d1a4a9318a7ef02a0fdedccb03f';
-const BASE_URL = 'https://api.themoviedb.org/3';
-const MOVIE_ID = 894205; /* example id */
+import { KEY, BASE_URL } from './key-url';
 
 export default class PopFilmsAPI {
   async fetchTrailer(movieId) {
