+++ src/js/fetch/key-url.js
@@ -1,0 +1,3 @@
+export const KEY = '3c2d3d1a4a9318a7ef02a0fdedccb03f';
+
+export const BASE_URL = 'https://api.themoviedb.org/3';
