+++ src/js/refs/filmoteka-refs.js
@@ -1,0 +1,8 @@
+export const refs = {
+  formEl: document.querySelector('#search-form'),
+  inputEl: document.querySelector('#search'),
+  filmGalleryContainer: document.querySelector('.film-container'),
+  paginationInHome: document.querySelector('.pagination'),
+  notificationEl: document.querySelector('.text-error'),
+  notificationPictureEl: document.querySelector('#notification-picture'),
+};
