+++ src/js/utils-js/error-function.js
@@ -1,0 +1,9 @@
+import { refs } from '../refs/filmoteka-refs';
+
+export function onFetchError(error) {
+  console.log(error.message);
+
+  refs.notificationEl.style.color = '#ff001b';
+  refs.notificationEl.textContent =
+    'Sorry, something going wrong... Please try again.';
+}
