+++ src/js/utils-js/settimeout.js
@@ -1,0 +1,7 @@
+import { refs } from '../refs/filmoteka-refs';
+
+export function setClearTextContentTimer() {
+  setTimeout(() => {
+    refs.notificationEl.textContent = '';
+  }, 3000);
+}
